fix: Change CategoryId type from long to int in CreateProductDto

+++ backend/Core/DTOs/Product/CreateProductDto.cs
@@ -14,7 +14,7 @@ public record class CreateProductDto
     public required decimal Price { get; init; }
 
     [Required]
-    public long CategoryId { get; init; }
+    public int CategoryId { get; init; }
 
     /// <summary>
     /// Display name or alt text for the product image
